+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 - DefaultAudioMixer
+- Dependency to TextMeshPro package
 
 ## [1.1.0] - 2020-02-16
 ### Added