+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - DefaultAudioMixer
 - Dependency to TextMeshPro package
 - AudioManagerData
+- AudioManager
 
 ## [1.1.0] - 2020-02-16
 ### Added